+++ src/main/java/online/kingdomkeys/kingdomkeys/capability/GlobalCapabilities.java
@@ -7,9 +7,7 @@ public class GlobalCapabilities implements IGlobalCapabilities {
 	@Override
 	public CompoundTag serializeNBT() {
 		CompoundTag storage = new CompoundTag();
-		storage.putInt("ticks_stopped", this.getStoppedTicks());
 		storage.putFloat("stop_dmg", this.getStopDamage());
-		storage.putInt("ticks_flat", this.getFlatTicks());
 		storage.putBoolean("castle_oblivion_marker", this.getCastleOblivionMarker());
 		storage.putInt("level", this.getLevel());
 		storage.putBoolean("is_ko", isKO);
@@ -19,9 +17,7 @@ public class GlobalCapabilities implements IGlobalCapabilities {
 	@Override
 	public void deserializeNBT(CompoundTag nbt) {
 		CompoundTag properties = (CompoundTag) nbt;
-		this.setStoppedTicks(properties.getInt("ticks_stopped"));
 		this.setStopDamage(properties.getFloat("stop_dmg"));
-		this.setFlatTicks(properties.getInt("ticks_flat"));
 		this.setCastleOblivionMarker(properties.getBoolean("castle_oblivion_marker"));
 		this.setLevel(properties.getInt("level"));
 		this.setKO(properties.getBoolean("is_ko"));
@@ -40,21 +36,6 @@ public class GlobalCapabilities implements IGlobalCapabilities {
 	@Override
 	public int getLevel() {
 		return level;
-	}
-	
-	@Override
-	public void setStoppedTicks(int time) {
-		this.timeStopped = time;
-	}
-
-	@Override
-	public int getStoppedTicks() {
-		return timeStopped;
-	}
-
-	@Override
-	public void subStoppedTicks(int time) {
-		this.timeStopped -= time;
 	}
 
 	@Override
@@ -81,21 +62,6 @@ public class GlobalCapabilities implements IGlobalCapabilities {
 	@Override
 	public String getStopCaster() {
 		return this.stopCaster;
-	}
-
-	@Override
-	public int getFlatTicks() {
-		return flatTicks;
-	}
-
-	@Override
-	public void setFlatTicks(int time) {
-		this.flatTicks = time;
-	}
-
-	@Override
-	public void subFlatTicks(int time) {
-		this.flatTicks -= time;
 	}
 
 	@Override