+++ src/main/java/online/kingdomkeys/kingdomkeys/capability/IGlobalCapabilities.java
@@ -4,10 +4,6 @@ import net.minecraft.nbt.CompoundTag;
 import net.minecraftforge.common.util.INBTSerializable;
 
 public interface IGlobalCapabilities extends INBTSerializable<CompoundTag> {
-	//Stop
-	int getStoppedTicks();
-	void setStoppedTicks(int time);
-	void subStoppedTicks(int time);
 	
 	//Stop damage
 	float getStopDamage();
@@ -15,12 +11,6 @@ public interface IGlobalCapabilities extends INBTSerializable<CompoundTag> {
 	void addDamage(float dmg);
 	void setStopCaster(String name);
 	String getStopCaster();
-	
-	
-	//Gravity Flat
-	int getFlatTicks();
-	void setFlatTicks(int time);
-	void subFlatTicks(int time);
 
 	//Castle Oblivion
 	boolean getCastleOblivionMarker();