+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/StopGui.java
@@ -5,6 +5,7 @@ import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
 import online.kingdomkeys.kingdomkeys.capability.ModCapabilities;
+import online.kingdomkeys.kingdomkeys.effects.ModMobEffects;
 import org.jetbrains.annotations.NotNull;
 import org.lwjgl.glfw.GLFW;
 
@@ -36,7 +37,7 @@ public class StopGui extends Screen {
 
 			if(ModCapabilities.getGlobal(minecraft.player) != null) {
 	            //InputConstants.grabOrReleaseMouse(this.minecraft.getWindow().getWindow(), 212993, 0, 0);
-				if(ModCapabilities.getGlobal(minecraft.player).getStoppedTicks() <= 0) {
+				if(minecraft.player.hasEffect(ModMobEffects.STOP.get())) {
 					onClose();
 				}