+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 **Added:**
 * Blizzard freeze effect, duration is longer for higher level magic and will stack duration for every hit
+* Aero, Gravity and Stop now work with mob effects